06-http-replay/requirements-dev.txt
pytest
python-dotenv[cli]
pytest-retry
pytest-vcr
